requirements.txt
django-autoslug==1.9.3
django-extensions==2.1.3
django-extra-views==0.12.0
django-guardian==1.4.9
django-parler==1.9.2
django-userena-ce==4.1.0
django-widget-tweaks==1.4.3
easy-thumbnails==2.5
html2text==2014.12.29
Pillow==5.3.0
pytz==2018.7
six==1.11.0
typing==3.6.6
